22/11/2562 5:44 Game update music

+++ Monster_Adventure/src/Stage.h
@@ -45,7 +45,7 @@ public:
 	virtual ~Stage();
 
 	//Accessor
-	const sf::Vector2f getStageSize();
+	const sf::Vector2i getStageSize();
 
 	const bool& getClear();
 	std::vector<Enemy*>& getEnemy();
